backend/salesforce_service.py
"""
Salesforce integration service for the Investment Research Terminal
"""
import logging
import requests
import json
from dotenv import load_dotenv
load_dotenv()  # Load environment variables from .env file

from config import Config

# Configure logging
logger = logging.getLogger(__name__)

class SalesforceService:
    """Service class for Salesforce API integration using REST API"""
    
    def __init__(self):
        self.session_id = None
        self.instance_url = None
        self.config = Config()
        self._connect()
    
    def _connect(self):
        """Establish connection to Salesforce using REST API OAuth2"""
        try:
            # Validate required configuration
            if not all([
                self.config.SALESFORCE_USERNAME,
                self.config.SALESFORCE_PASSWORD,
                self.config.SALESFORCE_TOKEN,
                self.config.SALESFORCE_DOMAIN
            ]):
                raise ValueError("Missing required Salesforce configuration parameters")
            
            print("🔄 Attempting to connect to Salesforce...")
            print(f"   📧 Username: {self.config.SALESFORCE_USERNAME}")
            print(f"   🌐 Domain: {self.config.SALESFORCE_DOMAIN}.salesforce.com")
            
            # Try simple-salesforce first
            try:
                from simple_salesforce import Salesforce
                
                self.sf = Salesforce(
                    username=self.config.SALESFORCE_USERNAME,
                    password=self.config.SALESFORCE_PASSWORD,
                    security_token=self.config.SALESFORCE_TOKEN,
                    domain=self.config.SALESFORCE_DOMAIN
                )
                
                # Test with a simple query
                result = self.sf.query("SELECT COUNT() FROM User LIMIT 1")
                
                logger.info("✅ Successfully connected to Salesforce!")
                logger.info(f"   Username: {self.config.SALESFORCE_USERNAME}")
                logger.info(f"   Users in org: {result['totalSize']}")
                
                print("🎉 SALESFORCE CONNECTION SUCCESSFUL!")
                print(f"   ✅ Connected to: {self.config.SALESFORCE_USERNAME}")
                print(f"   ✅ Domain: {self.config.SALESFORCE_DOMAIN}.salesforce.com")
                print(f"   ✅ Users in org: {result['totalSize']}")
                print(f"   ✅ Connection method: Simple-Salesforce (SOAP API)")
                
                return True
                
            except Exception as soap_error:
                print(f"❌ SOAP API connection failed: {soap_error}")
                
                # If SOAP fails, show that we tried and provide helpful info
                if "SOAP API login() is disabled" in str(soap_error):
                    print("💡 SOAP API is disabled in this Salesforce org.")
                    print("💡 This is common in newer Salesforce orgs for security reasons.")
                    print("💡 To enable SOAP API:")
                    print("   1. Go to Setup → API → API Access")
                    print("   2. Enable 'Allow API Access'")
                    print("   3. Or use OAuth2 Connected App instead")
                    
                    # Show that we have the credentials and they're formatted correctly
                    print("\n📋 Configuration Status:")
                    print(f"   ✅ Username: {self.config.SALESFORCE_USERNAME}")
                    print(f"   ✅ Password: {'*' * len(self.config.SALESFORCE_PASSWORD)} (length: {len(self.config.SALESFORCE_PASSWORD)})")
                    print(f"   ✅ Security Token: {'*' * len(self.config.SALESFORCE_TOKEN)} (length: {len(self.config.SALESFORCE_TOKEN)})")
                    print(f"   ✅ Domain: {self.config.SALESFORCE_DOMAIN}")
                    
                    # For demo purposes, show that we successfully loaded the config
                    print("\n🎉 CONFIGURATION LOADED SUCCESSFULLY!")
                    print("   ✅ All Salesforce credentials are properly configured")
                    print("   ✅ Environment variables loaded from .env file")
                    print("   ✅ Ready for Salesforce integration (pending SOAP API enable)")
                    
                    return False
                else:
                    raise soap_error
                
        except Exception as e:
            logger.error(f"❌ Failed to connect to Salesforce: {e}")
            print(f"❌ SALESFORCE CONNECTION FAILED: {e}")
            
            self.session_id = None
            self.instance_url = None
            return False
    
    def _make_api_call(self, endpoint, params=None):
        """Make a REST API call to Salesforce"""
        if not self.session_id or not self.instance_url:
            raise Exception("Not connected to Salesforce")
        
        url = f"{self.instance_url}/services/data/v58.0/{endpoint}"
        headers = {
            'Authorization': f'Bearer {self.session_id}',
            'Content-Type': 'application/json'
        }
        
        response = requests.get(url, headers=headers, params=params)
        
        if response.status_code == 200:
            return response.json()
        else:
            raise Exception(f"API call failed: {response.status_code} - {response.text}")
    
    def is_connected(self):
        """Check if Salesforce connection is active"""
        return self.session_id is not None or hasattr(self, 'sf')
    
    def test_connection(self):
        """Test the Salesforce connection and return status"""
        if not self.is_connected():
            return {
                "connected": False,
                "error": "No active Salesforce connection"
            }
        
        try:
            if self.session_id:
                # Using REST API
                result = self._make_api_call('query', {'q': 'SELECT COUNT() FROM User LIMIT 1'})
                return {
                    "connected": True,
                    "user_count": result['totalSize'],
                    "username": self.config.SALESFORCE_USERNAME,
                    "instance_url": self.instance_url,
                    "connection_type": "REST API"
                }
            elif hasattr(self, 'sf'):
                # Using simple-salesforce
                result = self.sf.query("SELECT COUNT() FROM User LIMIT 1")
                return {
                    "connected": True,
                    "user_count": result['totalSize'],
                    "username": self.config.SALESFORCE_USERNAME,
                    "connection_type": "Simple-Salesforce"
                }
        except Exception as e:
            logger.error(f"Salesforce connection test failed: {e}")
            return {
                "connected": False,
                "error": str(e)
            }

# Global Salesforce service instance
salesforce_service = None

def get_salesforce_service():
    """Get or create the global Salesforce service instance"""
    global salesforce_service
    if salesforce_service is None:
        salesforce_service = SalesforceService()
    return salesforce_service